requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
chess==1.11.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
python-chess==1.999
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
numpy==2.1.2
torch==2.5.1
tqdm==4.66.5
modal==0.63.41
huggingface-hub==0.24.6
datasets==2.20.0
python-dotenv==1.0.0
